baseline fat32

+++ dependency/isomorphic_drivers/src/block/ahci.rs
@@ -447,7 +447,7 @@ impl<P: Provider> AHCI<P> {
         }
     }
 
-    pub fn read_block(&mut self, block_id: u64, buf: &mut [u8]) -> usize {
+    pub fn read_block(&mut self, block_id: usize, buf: &mut [u8]) -> usize {
         // cfl=4
         self.cmd_list[0].flags = 4;
 
@@ -470,7 +470,7 @@ impl<P: Provider> AHCI<P> {
         len
     }
 
-    pub fn write_block(&mut self, block_id: u64, buf: &[u8]) -> usize {
+    pub fn write_block(&mut self, block_id: usize, buf: &[u8]) -> usize {
         // cfl=4
         self.cmd_list[0].flags = 4 | CommandHeaderFlags::WRITE.bits(); // device write
